requirements.txt
numpy
shapely
pillow
pytest
matplotlib
opencv-python